Fix time_t definition in pwd.lua
time_t is long on all major platforms, not int.

Fixes #2443

+++ src/lua/pwd.lua
@@ -6,7 +6,7 @@ local errno = require('errno')
 ffi.cdef[[
     typedef int uid_t;
     typedef int gid_t;
-    typedef int time_t;
+    typedef long time_t;
 ]]
 
 -- POSIX demands to have three fields in struct group:
